add pomFromGradleProperties API to base plugin
